[UBLE-62] refactor : 피드백 기반 리팩토링

+++ src/main/java/com/ureca/uble/domain/pin/repository/PinRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 public interface PinRepository extends JpaRepository<Pin, Long> {
 	void deleteByUser(User user);
 	List<Pin> findByUserIdOrderByIdAsc(Long userId);
-	long countByUserId(Long userId);
+	int countByUserId(Long userId);
 	Optional<Pin> findByIdAndUserId(Long pinId, Long userId);
 	void deleteByIdAndUserId(Long pinId, Long userId);
 }
